backend/tttn_no_park_zone/requirements.txt
ultralytics>=8.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
shapely>=2.0.0
opencv-python>=4.8.0
shapely>=2.0.0
pyyaml>=6.0.0
numpy>=1.24.0
Pillow>=9.0.0
